Module Odig_support.Opam

Lookup package opam metadata.

opam metadata

type t

The type for opam metadata.

val authors : t -> string list

authors i is the authors: field.

val bug_reports : t -> string list

bug_reports i is the bug-reports: field.

val depends : t -> string list

depends i is the depends: field.

val dev_repo : t -> string list

dev_repo i is the dev-repo: field.

val doc : t -> string list

doc i is the doc: field.

val homepage : t -> string list

homepage i is the homepage: field.

val license : t -> string list

license i is the license: field.

val maintainer : t -> string list

maintainer i is the maintainer: field.

val synopsis : t -> string

synopsis i is the synopsis: field.

val tags : t -> string list

info_tags i are the package's tags.

val version : t -> string

version i is the package's version.

Queries

val file : Pkg.t -> B00_std.Fpath.t option

file pkg is the opam file of package pkg (if any).

val query : Pkg.t list -> (Pkg.t * t) list

query pkgs queries the opam files associated to the given packages (if any).

Note. It is better to batch queries, opam show is quite slow (at least until v2.0.3).