Module Odig_support.Pkg_info

Gather package information

Gathers Doc_cobj, Opam and Docdir information about a package.

Package info

type t

The type for package information.

val doc_cobjs : t -> Doc_cobj.t list

doc_cobjs i are the documentation compilation objects of i.

val doc_dir : t -> Doc_dir.t

doc_dir i is the doc dir information of i.

val opam : t -> Opam.t

opam i is the opam information of i.

Uniform field access

type field = [

| `Authors

| `Changes_files

| `Depends

| `Doc_cobjs

| `Homepage

| `Issues

| `License

| `License_files

| `Maintainers

| `Odoc_assets

| `Odoc_pages

| `Online_doc

| `Readme_files

| `Repo

| `Synopsis

| `Tags

| `Version

]

The type for fields.

val field_names : (string * field) list

field_names associated a string name to each field.

val get : field -> t -> string list

get field i is the field field of i as a list of strings.

val pp : t B00_std.Fmt.t

pp formats all package information fields in alphabetic order.

Queries

val query : doc_dir:B00_std.Fpath.t -> Pkg.t list -> (Pkg.t * t) list

query ~doc_dir pkgs combines the result of Doc_cobj.of_pkg, Opam:query and Doc_dir.of_pkg.