Package b0 0.0.2-12-g2e7fa30e

WARNING. This package is unstable and work in progress, do not depend on it.

B0 describes software construction and deployments using modular and customizable definitions written in OCaml.

B0 is built on top of the B00 build library which provides arbitrary build abstraction with reliable, efficient incremental rebuilds.

Manuals

For B0 users.

The B0 manual is a conceptual overview of the B0 system and B0 files. It is recommended reading.

The build unit manual describes how to create your own build units.

The cmdlet manual describes how to define your own cmdlets.

The B00 manual describes the B00 build library.

The B0 release manual describes B0 support for releasing software.

For B0 users of specific tools.

The B0 opam manual describes the opam support to generate opam package files and publish them in repositories.

For B0 programmers.

The B0 driver development manual describes how to make your own drivers to operate on B0 files.

The TODO page

Old manual fragments and TODO which will eventually disapear.

The rationale of B0, written a long time ago but still relevant

B00

B00 is the build model used by B0. It can be used for its own good to devise domain specific build system – see for example the brzo or odig tools.

The b0.b00.std and b0.b00.kit libraries also provide sharp tools for quick "scripts" and command line tools even if you are not using b0.b00.

Library b0.b00.std

The b0.b00.std library has a few things that should be in the stdlib and a few others that should not be but are useful for B0 and B00 based programs.

B00_std Standard needs for B0 and B00 programs (and sometimes for any program but it's getting better).

Library b0.b00

The build library.

B000 Build kernel plumbing.

B000_conv Serialize and format B000 values.

B00 Build kernel

Library b0.b00.kit

Toolkit for working with B00 and B0. A few things do not depend on the b0.b00 library but are useful to have under your hand.

IO and file formats

B00_base64 Base64 codec.

B00_fexts Sets of file extension sets.

B00_findex File indexes.

B00_htmlg HTML generation.

B00_http World Wide Web interaction.

B00_lines Parse text lines.

B00_serialk_json JSON text support.

B00_serialk_sexp S-expression support.

B00_serialk_text UTF-8 text lexing tools.

B00_trace Trace build operations.

Dev tools and info

B00_cmark B00 cmark support.

B00_github GitHub interaction.

B00_jsoo B00 js_of_ocaml support.

B00_ocaml B00 ocaml support.

B00_odoc B0 odoc support.

B00_os Operating system and machine information.

B00_vcs Version control system (VCS) repositories.

B00_rsync rsync support.

UI fragments

B00_cli Command line user interface fragments.

Viewers and editors

B00_editor Editor interaction.

B00_pager Pager interaction.

B00_pdf_viewer PDF viewer interaction.

B00_www_browser Web browser interaction.

B0

B0 is the system for describing software construction and deployments via B0 files.

Library b0

B0_build B0 builds.

B0_def B0 definitions.

B0_dir B0_dir directory structured access.

B0_meta Metadata.

B0_cli Command line user interface fragments.

B0 definitions

B0_cmdlet Cmdlets.

B0_pack Build packs.

B0_unit Build units.

B0 files

Only needed if you want to process B0 files your own own way. See the B0 driver development manual.

B0_file B0 file source and expansion.

B0_driver B0 file drivers.

Library b0.kit

Toolkit for writing your B0 files.

B0_kit APIs to use in B0 files.

B0_jsoo js_of_ocaml B0 file support

B0_ocaml OCaml B0 file support.

B0_ocaml_eco OCaml eco-system support.

B0_opam B0 opam support.

B0_release B0 software release helpers.

B0_srcs Select source files.

Package info

authors

The b0 programmers

changes-files

CHANGES.md

depends

cmdliner

ocaml

ocamlbuild

ocamlfind

topkg

homepage

https://erratique.ch/software/b0

issues

https://github.com/b0-system/b0/issues

license

BSD2

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/b0/doc

readme-files

README.md

repo

git+https://erratique.ch/repos/b0.git

tags

build

dev

org:b0-system

org:erratique

version

0.0.2-12-g2e7fa30e