Package batteries

BatArray

BatAvlTree Internals of ISet and IMap, usable as generic tree library

BatBase64 Base64 codec.

BatBig_int Operations on arbitrary-precision integers.

BatBigarray Additional and modified functions for big arrays.

BatBitSet Efficient bit sets.

BatBool Operations on booleans

BatBounded Bounded values

BatBuffer Extensible string buffers.

BatBytes Byte sequence operations.

BatBytesCompat

BatCache

BatChar Operations on characters.

BatCharParser Parsing character strings.

BatComplex Additional and modified functions for complex numbers.

BatConcreteQueue

BatConcurrent Definition of concurrency primitives.

BatDeque Functional double-ended queues

BatDigest

BatDllist A mutable, imperative, circular, doubly linked list library

BatDynArray Dynamic arrays.

BatEither Either type.

BatEnum Enumeration over abstract collection of elements.

BatFile File manipulation.

BatFilename Operations on file names.

BatFingerTree This module implements a generic finger tree datastructure as described here: Finger Trees: A Simple General-purpose Data Structure http://www.soi.city.ac.uk/~ross/papers/FingerTree.pdf

BatFloat Operations on floating-point numbers.

BatFormat Pretty printing.

BatGc Memory management control and statistics; finalised values.

BatGenlex A generic lexical analyzer.

BatGlobal Mutable global variable.

BatHashcons Hash consing of data structures

BatHashtbl Extra functions over hashtables.

BatHeap Functional heaps over ordered types

BatIMap DIET Maps from integers, packed using ranges

BatIO High-order abstract I/O.

BatISet DIET : Discrete Interval Encoding Trees

BatInnerIO Core of the BatIO module.

BatInnerPervasives

BatInnerShuffle

BatInnerWeaktbl Weak hash table library for OCaml, with an interface compatible with the standard Hashtbl module.

BatInt Operations on integers.

BatInt32

BatInt64 64-bit integers.

BatInterfaces Common signatures for data structures.

BatLazyList Lazy lists of elements.

BatLexing Simple lexing using ocaml conventions

BatList Additional and modified functions for lists.

BatLog

BatLogger

BatMap

BatMarshal Marshaling of data structures.

BatMultiMap Polymorphic Multi-Map.

BatMultiPMap Polymorphic Multi-Map.

BatMutex Locks for mutual exclusion.

BatNativeint Processor-native integers.

BatNum Operation on arbitrary-precision numbers.

BatNumber A common interface for numbers.

BatOpaqueInnerSys

BatOptParse Modules for GNU getopt(3)-style command line parsing.

BatOption Functions for the option type.

BatOrd

BatParserCo A simple parser combinator library.

BatPathGen Filepath handling.

BatPervasives

BatPrintexc Facilities for printing exceptions.

BatPrintf Formatted output functions (also known as unparsing).

BatQueue First-in first-out queues.

BatRMutex Reentrant Mutexes

BatRandom Pseudo-random number generators (PRNG).

BatRef Operations on references.

BatRefList Reference on lists.

BatResult Monadic results of computations that can raise exceptions

BatReturn Local exceptions/labels/goto/return.

BatScanf Formatted input functions.

BatSeq Sequence of elements

BatSet Sets over ordered types.

BatSplay Maps over ordered types based on splay trees.

BatStack Last-in first-out stacks.

BatStream Streams and stream parsers

BatString String operations.

BatSubstring

BatSys System interface.

BatText Heavyweight strings ("ropes")

BatTuple Tuples.

BatUChar Unicode characters.

BatUTF8 UTF-8 encoded Unicode strings. The type is normal string.

BatUnit Operations on unit.

BatUnix Low-level interface to the operating system (both Unix and Windows).

BatUref Unifiable references using destructive union-find

BatVect Extensible vectors with constant-time append/prepend.

Batteries

BatteriesConfig

BatteriesExceptionless

BatteriesHelp Tools for reading the documentation from the toplevel

BatteriesPrint

BatteriesThread

Extlib

Package info