Package brzo 79d316a5

Brzo is a build tool to quickly turn source files of various languages into executable programs and documents.

Consult the manual and its tutorial introduction.

Quick start

A few invocations to get you started.

> touch BRZO
> cat > echo.ml <<EOCAML
let echo oc ss = output_string oc (String.concat " " ss ^ "\n")
let () = echo stdout (List.tl (Array.to_list Sys.argv))
EOCAML
> brzo -- 'Quick!'
'Quick!'
> brzo ocaml --html -- 'Quick!'   # See your browser console.
> brzo --doc

More in the manual.

API

This is an unstable API subject to change even between minor versions of the tool. Use at your own risk.

Brzo Basic definitions and domain helpers.

Brzo_domain Brzo domains.

Brzo_domain_list Available domains.

Brzo_tie_conf Ties the configuration knot.

Brzo_outcome Outcomes

Brzo_b0_c C B0 helpers.

Brzo_b0_js_of_ocaml B0 js_of_ocaml support.

Brzo_b0_latex LaTeX domain B0 helpers.

Brzo_b0_ocaml B0 ocaml support.

Brzo_b0_ocamlpath

Brzo_b0_opam Opam b0 helpers.

Brzo_c C domain.

Brzo_cmark CommonMark domain.

Brzo_latex LaTeX domain.

Brzo_read_cmi

Brzo_ocaml_conf Brzo OCaml domain configuration.

Brzo_ocaml_cmi Compiled interfaces.

Brzo_ocaml_be OCaml domain best-effort helpers.

Brzo_ocaml OCaml domain.

Package info

authors

The brzo programmers

changes-files

CHANGES.md

depends

b0

cmdliner

ocaml

ocamlbuild

ocamlfind

topkg

homepage

https://erratique.ch/software/brzo

issues

https://github.com/b0-system/brzo/issues

license

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/brzo/doc

readme-files

README.md

repo

git+https://erratique.ch/repos/brzo.git

tags

build

dev

org:b0-system

org:erratique

version

79d316a5