Package ocamlbuild 0.14.0

Ocamlbuild_executor

Ocamlbuild_pack

Ocamlbuild_plugin

Ocamlbuild_unix_plugin

Signatures This module contains all module signatures that the user could use to build an ocamlbuild plugin.

Package info

authors

Berke Durak

Nicolas Pouillard

changes-files

Changes

depends

ocaml

homepage

https://github.com/ocaml/ocamlbuild/

issues

https://github.com/ocaml/ocamlbuild/issues

license

LGPL-2 with OCaml linking exception

license-files

LICENSE

maintainers

Gabriel Scherer <gabriel.scherer@gmail.com>

online-doc

https://github.com/ocaml/ocamlbuild/blob/master/manual/manual.adoc

readme-files

Readme.md

repo

git+https://github.com/ocaml/ocamlbuild.git

version

0.14.0