Package opam-core 2.1.0~beta4

OpamACL C auxiliary function used for POSIX 1003.1e DRAFT 17 permission checking.

OpamCached

OpamCompat

OpamConsole Console output, ANSI color, logging and user querying

OpamCoreConfig Configuration options for the core lib (record, global reference and setter)

OpamDirTrack

OpamFilename Higher level file and directory name manipulation AND file operations, wrappers on OpamSystem using the filename type

OpamHash

OpamJson Json encoder; only needed for some debug options

OpamParallel Parallel execution of jobs following a directed graph

OpamProcess Process and job handling, with logs, termination status, etc.

OpamSHA Pure OCaml implementation of SHA256/512 hashing functions. The hash is returned as an hex string.

OpamStd Generic stdlib functions (String, List, Option, Sys submodules...)

OpamStubs OS-specific functions requiring C code on at least one platform.

OpamStubsTypes Types for C stubs modules.

OpamSystem Bindings of lots of filesystem and system operations

OpamUrl URL parsing and printing, with support for our different backends

OpamVersion (generated) Current OPAM version

OpamVersionCompare Version comparison function used throughout. From the Dose suite.

Package info

authors

Anil Madhavapeddy <anil@recoil.org>

Fabrice Le Fessant <Fabrice.Le_fessant@inria.fr>

Frederic Tuong <tuong@users.gforge.inria.fr>

Guillem Rieu <guillem.rieu@ocamlpro.com>

Louis Gesbert <louis.gesbert@ocamlpro.com>

Raja Boujbel <raja.boujbel@ocamlpro.com>

Ralf Treinen <ralf.treinen@pps.jussieu.fr>

Roberto Di Cosmo <roberto@dicosmo.org>

Thomas Gazagnaire <thomas@gazagnaire.org>

Vincent Bernardoff <vb@luminar.eu.org>

changes-files

CHANGES

depends

base-bigarray

base-unix

cppo

dune

ocaml

ocamlgraph

re

homepage

https://opam.ocaml.org/

issues

https://github.com/ocaml/opam/issues

license-files

LICENSE

maintainers

opam-devel@lists.ocaml.org

readme-files

README.md

repo

git+https://github.com/ocaml/opam.git

version

2.1.0~beta4