Package opam-format 2.1.0~beta4

OpamFile Handles all OPAM file formats as record types and submodules, conversion to and from syntax

OpamFilter Formulas on variables, as used in opam files build scripts

OpamFormat OPAM files syntax and conversion tools

OpamFormatConfig Configuration options for the format lib (record, global reference and setter)

OpamFormula Formulas on packages, opt. with sub-formulas on versions, and conversion functions

OpamInterpLexer OPAM format variable interpolation processor

OpamLineLexer A simple lexer to list of lines, which are lists of words

OpamPackage The package type, and package name type (name+version, values often called "nv" in the code)

OpamPath Defines the file hierarchy in ~/.opam

OpamPp Generic bidirectional transformation toolbox for parsing/printing

OpamRepositoryName The type for repository names

OpamSwitch The type for switch names

OpamSysPkg

OpamTypes Definitions of many types used throughout

OpamTypesBase Helper functions on the base types (from OpamTypes)

OpamVariable OPAM variables with scope (global or module), used in "opam" package definition files in "filters"

Package info

authors

Anil Madhavapeddy <anil@recoil.org>

Fabrice Le Fessant <Fabrice.Le_fessant@inria.fr>

Frederic Tuong <tuong@users.gforge.inria.fr>

Guillem Rieu <guillem.rieu@ocamlpro.com>

Louis Gesbert <louis.gesbert@ocamlpro.com>

Raja Boujbel <raja.boujbel@ocamlpro.com>

Ralf Treinen <ralf.treinen@pps.jussieu.fr>

Roberto Di Cosmo <roberto@dicosmo.org>

Thomas Gazagnaire <thomas@gazagnaire.org>

Vincent Bernardoff <vb@luminar.eu.org>

changes-files

CHANGES

depends

dune

ocaml

opam-core

opam-file-format

homepage

https://opam.ocaml.org/

issues

https://github.com/ocaml/opam/issues

license-files

LICENSE

maintainers

opam-devel@lists.ocaml.org

readme-files

README.md

repo

git+https://github.com/ocaml/opam.git

version

2.1.0~beta4