Package uri 4.1.0

Uri Uniform Resource Identifier handling that is RFC3986-compliant.

Uri_services Module that maps the more common IANA well-known ports to and from their textual service names. To save memory, this module does not contains only the more common services. Use the Uri_services_full module if you need the complete list.

Uri_services_full Module that maps the IANA well-known ports to and from their textual service names

Package info

authors

Anil Madhavapeddy

David Sheets

Rudi Grinberg

changes-files

CHANGES.md

depends

angstrom

dune

ocaml

ounit

ppx_sexp_conv

stringext

homepage

https://github.com/mirage/ocaml-uri

issues

https://github.com/mirage/ocaml-uri/issues

license

ISC

license-files

LICENSE.md

maintainers

anil@recoil.org

online-doc

https://mirage.github.io/ocaml-uri/

readme-files

README.md

repo

git+https://github.com/mirage/ocaml-uri.git

tags

org:mirage

org:xapi-project

uri

url

version

4.1.0