Package yojson 1.7.0

Yojson The Yojson library provides runtime functions for reading and writing JSON data from OCaml. It addresses a few shortcomings of its predecessor json-wheel and is about twice as fast (2.7x reading, 1.3x writing; results may vary). The design goals of Yojson are the following:

Yojson_biniou Conversions between JSON and biniou

Package info

authors

Martin Jambon

changes-files

CHANGES.md

depends

alcotest

biniou

cppo

dune

easy-format

ocaml

homepage

https://github.com/ocaml-community/yojson

issues

https://github.com/ocaml-community/yojson/issues

license-files

LICENSE.md

maintainers

martin@mjambon.com

online-doc

https://ocaml-community.github.io/yojson/

readme-files

README.md

repo

git+https://github.com/ocaml-community/yojson.git

version

1.7.0