Module Odig_support.Env

Odig environment variables.

Environment variables

val b0_cache_dir : string

b0_cache_dir is the environment variable that can be used to define the odig b0 cache directory.

val b0_log_file : string

b0_log_file is the environment variable that can be used to define the odig b0 log_file.

val cache_dir : string

cache_dir is the environment variable that can be used to define the odig cache directory.

val color : string

color is the variable used to specify TTY styling.

val doc_dir : string

doc_dir is the environment variable that can be used to define a doc dir.

val lib_dir : string

lib_dir is the environment variable that can be used to define a lib dir.

val odoc_theme : string

odoc_theme is the environment variable that can be used to define the default odoc theme.

val share_dir : string

share_dir_env is the environment variable that can be used to define a share dir.

val verbosity : string

verbosity is the variable used to specify log verbosity.